+++ R/model-xgboost.R
@@ -36,7 +36,7 @@ get_xgb_path <- function(row_id, tree) {
 
 get_xgb_tree <- function(tree) {
   paths <- seq_len(nrow(tree))[tree[, "Feature"] == "Leaf"]
-  map(
+  x <- map(
     paths,
     ~ {
       list(
@@ -45,32 +45,40 @@ get_xgb_tree <- function(tree) {
       )
     }
   )
+  x
 }
 
 get_xgb_trees <- function(model) {
-  UseMethod("get_xgb_trees", model)
-}
-
-get_xgb_trees.xgb.Booster <- function(model) {
-  xgb_dump_text_with_stats <- xgboost::xgb.dump(model, dump_format = "text", with_stats = TRUE)
+  xd <- xgboost::xgb.dump(
+    model = model,
+    dump_format = "text",
+    with_stats = TRUE
+  )
   feature_names <- model$feature_names
-  get_xgb_trees.character(xgb_dump_text_with_stats, feature_names)
+  get_xgb_trees_character(xd, feature_names)
 }
 
-get_xgb_trees.character <- function(xgb_dump_text_with_stats, feature_names) {
+get_xgb_trees_character <- function(xd, feature_names) {
   feature_names_tbl <- data.frame(
     Feature = as.character(0:(length(feature_names) - 1)),
     feature_name = feature_names,
     stringsAsFactors = FALSE
   )
-  trees <- xgboost::xgb.model.dt.tree(text = xgb_dump_text_with_stats)
+  trees <- xgboost::xgb.model.dt.tree(text = xd)
   trees <- as.data.frame(trees)
   trees$original_order <- 1:nrow(trees)
   trees <- merge(trees, feature_names_tbl, by = "Feature", all.x = TRUE)
   trees <- trees[order(trees$original_order), !names(trees) %in% "original_order"]
-  trees[, c("Yes", "No", "Missing")] <- lapply(trees[, c("Yes", "No", "Missing")], function(x) sub("^.*-", "", x))
-  trees[, c("Yes", "No", "Missing")] <- lapply(trees[, c("Yes", "No", "Missing")], function(x) as.integer(x) + 1)
-  purrr::map(split(trees, trees$Tree), get_xgb_tree)
+  trees[, c("Yes", "No", "Missing")] <-
+    lapply(trees[, c("Yes", "No", "Missing")], function(x) sub("^.*-", "", x))
+  trees[, c("Yes", "No", "Missing")] <-
+    lapply(trees[, c("Yes", "No", "Missing")], function(x) as.integer(x) + 1)
+
+  trees_split <- split(trees, trees$Tree)
+  trees_rows <- purrr::map_dbl(trees_split, nrow)
+  trees_filtered <- trees_split[trees_rows > 1]
+  
+  purrr::map(trees_filtered, get_xgb_tree)
 }
 
 #' @export
@@ -97,14 +105,26 @@ get_xgb_case <- function(path, prediction) {
   cl <- map(
     path,
     ~ {
-      if (.x$op == "less" & .x$missing) i <- expr((!!sym(.x$col) >= !!as.numeric(.x$val) | is.na(!!sym(.x$col))))
-      if (.x$op == "more-equal" & .x$missing) i <- expr((!!sym(.x$col) < !!as.numeric(.x$val) | is.na(!!sym(.x$col))))
-      if (.x$op == "less" & !.x$missing) i <- expr(!!sym(.x$col) >= !!as.numeric(.x$val))
-      if (.x$op == "more-equal" & !.x$missing) i <- expr(!!sym(.x$col) < !!as.numeric(.x$val))
+      if (.x$op == "less" & .x$missing) {
+        i <- expr((!!sym(.x$col) >= !!as.numeric(.x$val) | is.na(!!sym(.x$col))))
+      }
+      if (.x$op == "more-equal" & .x$missing) {
+        i <- expr((!!sym(.x$col) < !!as.numeric(.x$val) | is.na(!!sym(.x$col))))
+      }
+      if (.x$op == "less" & !.x$missing) {
+        i <- expr(!!sym(.x$col) >= !!as.numeric(.x$val))
+      }
+      if (.x$op == "more-equal" & !.x$missing) {
+        i <- expr(!!sym(.x$col) < !!as.numeric(.x$val))
+      }
       i
     }
   )
-  cl <- if (length(cl) > 0) reduce(cl, function(x, y) expr(!!x & !!y)) else TRUE
+  cl <- if (length(cl) > 0) {
+    reduce(cl, function(x, y) expr(!!x & !!y))
+  } else {
+    TRUE
+  }
   expr(!!cl ~ !!prediction)
 }
 
@@ -132,18 +152,26 @@ build_fit_formula_xgb <- function(parsedmodel) {
   if (is.null(objective)) {
     assigned <- 1
     f <- expr(!!f + !!base_score)
-    warning("If the objective is a custom function, please explicitly apply it to the output.")
-  } else if (objective %in% c("reg:squarederror")) {
+    warning(
+      paste(
+        "If the objective is a custom function, please",
+        "explicitly apply it to the output."
+      )
+    )
+  } else if (objective %in% c("reg:squarederror", "binary:logitraw")) {
     assigned <- 1
     f <- expr(!!f + !!base_score)
-  } else if (objective %in% c("binary:logitraw")) {
-    assigned <- 1
   } else if (objective %in% c("binary:logistic", "reg:logistic")) {
     assigned <- 1
     f <- expr(1 - 1 / (1 + exp(!!f + log(!!base_score / (1 - !!base_score)))))
   }
   if (assigned == 0) {
-    stop("Only objectives 'binary:logistic', 'reg:squarederror', 'reg:logistic', 'binary:logitraw' are supported yet.")
+    stop(
+      paste0(
+        "Only objectives 'binary:logistic', 'reg:squarederror',",
+        "'reg:logistic', 'binary:logitraw' are supported yet."
+      )
+    )
   }
   f
 }