+++ NEWS.md
@@ -1,5 +1,9 @@
 # tidypredict (development version)
 
+- Addresses issues with XGBoost models
+
+- Improvements to XGBoosts tests
+
 # tidypredict 0.4.9
 
 - Fixes issue handling GLM Binomial earth models (#97)